+++ example/android/app/src/main/kotlin/com/example/flek_select_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.flek_select_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()